+++ CHANGES.md
@@ -1,4 +1,5 @@
 #3.3.0.beta2 2013-??-??
+ - 2013-08-30 Fixed bug#[9670](http://bugs.otrs.org/show_bug.cgi?id=9670) - Ticket Number is removed wrongly from Article Subject if it contains a space.
  - 2013-08-30 Fixed bug#[9684](http://bugs.otrs.org/show_bug.cgi?id=9684) - Bulk select checkbox in AgentLinkObject does not work anymore.
  - 2013-08-30 Fixed bug#[9663](http://bugs.otrs.org/show_bug.cgi?id=9663) - Kernel\System\SysConfig.pm error message incomplete.
  - 2013-08-29 Fixed bug#[9431](http://bugs.otrs.org/show_bug.cgi?id=9431) - Terminology used for customer user contact.